+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Format new microerror@0.2.x JSON formatted error.
+
 ## [0.1.0] - 2020-07-01
 
 ### Changed